+++ CHANGELOG.md
@@ -15,6 +15,12 @@ All notable changes to this project will be documented here.
 
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### ✨ Enhancements
+
+- Create `useLocalStorage` hook.
+
 ## 1.0.0 (May 13, 2020)
 
 ### ✨ Enhancements